+++ ocdev/plugins/startapp/templates/7/app/tests/unit/controller/PageControllerTest.php
@@ -18,16 +18,15 @@ class PageControllerTest extends \PHPUnit_Framework_TestCase {
 	public function setUp () {
 		$app = new Application();
 		$this->container = $app->getContainer();
+		$this->container->registerService('Request', function($c) {
+			return $this->getMockBuilder('\OCP\IRequest')->getMock();
+		});
+		$this->container->registerParameter('UserId', 'john');
 	}
 
 
 	public function testIndex () {
-		// swap out request
-		$this->container['Request'] = $this->getMockBuilder('\OCP\IRequest')
-			->getMock();
-		$this->container['UserId'] = 'john';
-
-		$result = $this->container['PageController']->index();
+		$result = $this->container->query('PageController')->index();
 
 		$this->assertEquals(array('user' => 'john'), $result->getParams());
 		$this->assertEquals('main', $result->getTemplateName());
@@ -36,7 +35,7 @@ class PageControllerTest extends \PHPUnit_Framework_TestCase {
 
 
 	public function testEcho () {
-		$result = $this->container['PageController']->doEcho('hi');
+		$result = $this->container->query('PageController')->doEcho('hi');
 
 		$this->assertEquals(array('echo' => 'hi'), $result);
 	}